all hooks refactored out; working
